requirements.txt
boto3==1.9.215
botocore==1.12.215
docutils==0.15.2
jmespath==0.9.4
python-dateutil==2.8.0
PyYAML==5.1.2
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
